refactor(pre-tool-use): reorganize hooks and introduce helper modules
- Moved intent injection and validation logic into dedicated helper modules for better separation of concerns.
- Introduced `buildCtxBlock` and `buildMapBlock` functions to enhance context and map block construction.
- Updated `guardStructure` and `validatePreToolUse` functions to improve clarity and maintainability.
- Removed deprecated code and streamlined exports in the `pre-tool-use` module.
