+++ packages/filid/src/hooks/pre-tool-use/helpers/intent-injector/INTENT.md
