+++ packages/filid/src/hooks/pre-tool-use/helpers/intent-injector/utils/build-ctx-block.ts
