+++ packages/filid/src/hooks/pre-tool-use/helpers/intent-injector/utils/build-map-block.ts
