+++ packages/filid/src/hooks/pre-tool-use/helpers/intent-injector/utils/compress-paths.ts
