+++ packages/filid/src/hooks/pre-tool-use/helpers/pre-tool-validator/INTENT.md
