+++ packages/filid/src/hooks/pre-tool-use/helpers/pre-tool-validator/pre-tool-validator.ts
@@ -4,10 +4,10 @@ import * as path from 'node:path';
 import {
   validateDetailMd,
   validateIntentMd,
-} from '../../core/rules/document-validator/document-validator.js';
-import type { HookOutput, PreToolUseInput } from '../../types/hooks.js';
-import { isDetailMd, isIntentMd } from '../shared/shared.js';
-import { validateCwd } from '../utils/validate-cwd.js';
+} from '../../../../core/rules/document-validator/document-validator.js';
+import type { HookOutput, PreToolUseInput } from '../../../../types/hooks.js';
+import { isDetailMd, isIntentMd } from '../../../shared/shared.js';
+import { validateCwd } from '../../../utils/validate-cwd.js';
 
 const INTENT_MD_LINE_LIMIT = 50;
