+++ packages/filid/src/hooks/pre-tool-use/helpers/structure-guard/INTENT.md
