+++ packages/filid/src/hooks/pre-tool-use/helpers/structure-guard/index.ts
